% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compile.r
\name{compile_cr}
\alias{compile_cr}
\title{Compilation du compte rendu pdf}
\usage{
compile_cr(cfg_file = "config.r", encoding = "ISO8859-1", quiet = FALSE,
  recursive = TRUE, clean = TRUE, quiet_knit = TRUE)
}
\arguments{
\item{cfg_file}{cfg_file string. ficher R de configuration}

\item{encoding}{string. encodage du fichier cfg_file}

\item{quiet}{logical. desactive les sorties textuelles}

\item{recursive}{logical. voir \code{\link{dir.create}}}

\item{clean}{logical. active la supression des fichiers temporaires}

\item{quiet_knit}{logical. desactive les sorties textuelles de \code{\link[knitr]{knit2pdf}}}
}
\value{
The name of the output pdf file
}
\description{
Compilation du compte rendu pdf
}

