% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/export.r
\name{export_xl}
\alias{export_xl}
\title{Export data.frame to Excel file}
\usage{
export_xl(df, xlfile_out, open = TRUE)
}
\arguments{
\item{df}{data.frame}

\item{xlfile_out}{string. path to the output excel file}

\item{open}{logical. activate opening the file in Excel/LibreOffice/OpenOffice}
}
\description{
Export data.frame to Excel file
}

