% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lire.r
\name{lire_xl}
\alias{lire_xl}
\title{Read excel file}
\usage{
lire_xl(xlfile, sheets = 1, sheets_opt = "PHOTO", col_dates = c("DATE",
  "ECHEANCE"), origin = "1899-12-30")
}
\arguments{
\item{xlfile}{string. path to the file}

\item{sheets}{character or integer vector. mandatory sheets to be read}

\item{sheets_opt}{character or integer vector. optional sheets to be read}

\item{col_dates}{character vector. date column names}

\item{origin}{a Date object, or something which can be coerced by
  \code{as.Date(origin)} to such an object.}
}
\description{
Read excel file
}

