% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prepare.r
\name{open_fileman}
\alias{open_fileman}
\title{Open file manager to specified location}
\usage{
open_fileman(path, fileman = NULL)
}
\arguments{
\item{path}{string. path to the target location}

\item{fileman}{string. name of the file manager application}
}
\description{
Open file manager to specified location
}

