% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compile.r
\name{open_pdf}
\alias{open_pdf}
\title{Open pdf file}
\usage{
open_pdf(file, viewer = NULL, ...)
}
\arguments{
\item{file}{string. file to open}

\item{viewer}{string. name of the pdf viewer application (unix only)}

\item{...}{further arguments to be passed to \code{\link{system}}}
}
\description{
Open pdf file
}
\seealso{
\code{\link{system}}, \code{\link{shell}} (windows)
}

