% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare.r
\name{prepare_cr}
\alias{prepare_cr}
\title{Preparation du compte rendu}
\usage{
prepare_cr(cfg_file = "config.r", encoding = "ISO8859-1", quiet = FALSE,
  recursive = TRUE)
}
\arguments{
\item{cfg_file}{string. ficher R de configuration}

\item{encoding}{string. encodage du fichier cfg_file}

\item{quiet}{logical. desactive les sorties textuelles}

\item{recursive}{logical. voir \code{\link{dir.create}}}
}
\value{
\code{NULL}
}
\description{
Preparation du compte rendu
}

