% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/export.r
\name{prepare_photo}
\alias{prepare_photo}
\title{Prepare photo sheet to be completed}
\usage{
prepare_photo(photo_files)
}
\arguments{
\item{photo_files}{character vector. paths of the photo files}
}
\description{
Prepare photo sheet to be completed
}

