% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prepare.r
\name{prepare_photos}
\alias{prepare_photos}
\title{Prepare photo sheet to be completed}
\usage{
prepare_photos(photo_files = list.files(".", pattern =
  ".*\\\\.(jpg|jpeg|JPG|JPEG|png|PNG)"))
}
\arguments{
\item{photo_files}{character vector. paths of the photo files}
}
\description{
Prepare photo sheet to be completed
}
\keyword{internal}

