% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.R
\name{print_classe}
\alias{print_classe}
\title{Print object depending on its class}
\usage{
print_classe(cle, legende, format_fun, cle_var = "CLE",
  classe_var = "CLASSE")
}
\arguments{
\item{cle}{string. cle}

\item{legende}{data.frame}

\item{format_fun}{list. formatting functions}

\item{cle_var}{string. key column name}

\item{classe_var}{string. class column name}
}
\description{
Print object depending on its class
}

