% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.r
\name{print_photos}
\alias{print_photos}
\title{Print photos to latex}
\usage{
print_photos(photos,
  col_format = "|>{\\\\centering}m{.55\\\\linewidth}|m{.4\\\\linewidth}|",
  width = "\\\\linewidth", height = ".3\\\\textheight", ...)
}
\arguments{
\item{photos}{data.frame}

\item{col_format}{string. column format}

\item{width}{string. photo max width}

\item{height}{string. photo max height}

\item{...}{further arguments to be passed to print_table}
}
\description{
Print photos to latex
}

