% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.r
\name{print_table}
\alias{print_table}
\title{Print table to latex}
\usage{
print_table(df = data.frame(), envir = c("longtable", "tabular",
  "tabularx"), col_format = "", rowcolors = NULL,
  rowcol_head = "lightgray", hline = TRUE,
  format_head = "\\\\centering\\\\bf", replace_na = function(x)
  ifelse(is.na(x), "", x), replace_nl = function(x) gsub("\\n",
  "\\\\\\\\newline ", x), header = colnames(df), grid_col = "lightgray",
  vspace_before = "-.8em", vspace_after = "-.8em")
}
\arguments{
\item{df}{data.frame}

\item{envir}{string. table environment to be used in latex}

\item{col_format}{string. column format}

\item{rowcolors}{NULL or list with fields \code{begin} (integer. starting row including header)
\code{even} (string. even rows color) and \code{odd} (string. odd rows color)}

\item{rowcol_head}{string. row color of the columns header}

\item{hline}{logical. activate horizontal lines}

\item{format_head}{string. header columns format}

\item{replace_na}{function or NULL. replace NA}

\item{replace_nl}{function or NULL. replace newlines}

\item{header}{character vector or NULL. columns header text}

\item{grid_col}{string. color of the grid}

\item{vspace_before}{string or NULL. height of vertical space before the table}

\item{vspace_after}{string or NULL. height of vertical space after the table}
}
\description{
Print table to latex
}
\examples{
print_table(mtcars)
}

