% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.R
\name{print_taches}
\alias{print_taches}
\title{Print tasks to latex}
\usage{
print_taches(taches, legende, section = "orga",
  col_format = c("|p{11cm}|>{\\\\centering}p{2cm}|>{\\\\centering}p{2cm}|"),
  rowcol_head = "lightgray", header = c("TACHE", "ECHEANCE", "ETAT"),
  cle_var = "CLE", classe_var = "CLASSE", format_fun, ...)
}
\arguments{
\item{taches}{data.frame}

\item{legende}{data.frame}

\item{section}{string. section}

\item{col_format}{string. column format}

\item{rowcol_head}{row color of the header}

\item{header}{character vector. header}

\item{cle_var}{string. key column name}

\item{classe_var}{string. class column name}

\item{format_fun}{list. formatting functions}

\item{...}{further arguments to be passed to print_tab}
}
\description{
Print tasks to latex
}

