% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/compile.r
\name{resize_photos}
\alias{resize_photos}
\title{Resize photos}
\usage{
resize_photos(photo_files = list.files(".", pattern =
  ".*\\\\.(jpg|jpeg|JPG|JPEG|png|PNG)"), out_dir = tempdir(),
  max_width = 340, max_height = 340, quality = 95, quiet = FALSE)
}
\arguments{
\item{photo_files}{character vector. paths of the photo files}

\item{out_dir}{string. output directory of the resized photos}

\item{max_width}{maximum photo width}

\item{max_height}{maximum photo height}

\item{quality}{the 'quality' of the output JPEG image, as a percentage.}

\item{quiet}{logical. disable printing text output to the console}
}
\description{
Resize photos
}

