% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/export.r
\name{write_xl}
\alias{write_xl}
\title{Write data.frame to Excel file}
\usage{
write_xl(df, xlfile_out, open = FALSE, zip_path = "C:\\\\Rtools\\\\bin")
}
\arguments{
\item{df}{data.frame or list of data.frames}

\item{xlfile_out}{string. path to the output excel file}

\item{open}{logical. activate opening the file in Excel/LibreOffice/OpenOffice}

\item{zip_path}{string. path containing zip executable on windows}
}
\description{
Write data.frame to Excel file
}

