% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.r
\name{print_classe}
\alias{print_classe}
\title{Print object depending on its class}
\usage{
print_classe(cle, legende, format_fun)
}
\arguments{
\item{cle}{string. cle}

\item{legende}{data.frame}

\item{format_fun}{list. formatting functions}
}
\description{
Print object depending on its class
}

