% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.R
\name{print_photos}
\alias{print_photos}
\title{print photos to latex}
\usage{
print_photos(photos, col_format = "|>{\\\\centering}m{10cm}|m{7cm}|",
  width = "10cm", height = "5cm")
}
\arguments{
\item{photos}{data.frame}

\item{col_format}{string. column format}

\item{width}{string. photo max width}

\item{height}{string. photo max height}
}
\description{
print photos to latex
}

