% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.R
\name{print_plans}
\alias{print_plans}
\title{Print plans to latex}
\usage{
print_plans(plans, legende,
  col_format = c("|p{9cm}|>{\\\\centering}p{2cm}|>{\\\\centering}p{2cm}|>{\\\\centering}p{2cm}|"),
  rowcol_head = "lightgray", header = c("PLAN", "N°", "INDICE", "DATE"),
  cle_var = "CLE", classe_var = "CLASSE", ...)
}
\arguments{
\item{plans}{data.frame}

\item{legende}{data.frame}

\item{col_format}{string. column format}

\item{rowcol_head}{row color of the header}

\item{header}{character vector. header}

\item{cle_var}{string. key column name}

\item{classe_var}{string. class column name}

\item{...}{further arguments to be passed to print_tab}
}
\description{
Print plans to latex
}

