% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.R
\name{print_plans}
\alias{print_plans}
\title{Print plans to latex}
\usage{
print_plans(plans, legende, col_format, rowcol_head = "lightgray",
  header = c("PLAN", "NUM", "INDICE", "DATE"), format_fun, ...)
}
\arguments{
\item{plans}{data.frame}

\item{legende}{data.frame}

\item{col_format}{string. column format}

\item{rowcol_head}{row color of the header}

\item{header}{character vector. header}

\item{format_fun}{list. formatting functions}

\item{...}{further arguments to be passed to print_table}
}
\description{
Print plans to latex
}

