% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.R
\name{print_tab}
\alias{print_tab}
\title{Print table to latex}
\usage{
print_tab(df, col_format = "", rowcol = NULL, hline = TRUE,
  replace_na = function(x) ifelse(is.na(x), "", x), replace_nl = function(x)
  gsub("\\n", " \\\\\\\\newline ", x), header = NULL)
}
\arguments{
\item{df}{data.frame}

\item{col_format}{string. column format}

\item{rowcol}{string. row color}

\item{hline}{logical. activate horizontal lines}

\item{replace_na}{function or NULL}

\item{header}{character vector or NULL. header}
}
\description{
Print table to latex
}
\examples{
print_tab(mtcars)
}

