% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print.R
\name{print_taches}
\alias{print_taches}
\title{Print tasks to latex}
\usage{
print_taches(taches, legende, section = "orga", col_type = c("p{11cm}",
  "p{2cm}", "p{2cm}"), col_sep = "|", col_form = c("", ">{\\\\centering}",
  ">{\\\\centering}"), col_form_head = c(">{\\\\centering\\\\bf}",
  ">{\\\\centering\\\\bf}", ">{\\\\centering\\\\bf}"),
  rowcol_head = "lightgray", header, cle_var = "CLE",
  classe_var = "CLASSE", format_fun)
}
\arguments{
\item{taches}{data.frame}

\item{legende}{data.frame}

\item{section}{string. section}

\item{col_type}{character vector. column types}

\item{col_sep}{string. column separator}

\item{col_form}{character vector. column format}

\item{col_form_head}{character vector. header column format}

\item{rowcol_head}{row color of the header}

\item{header}{character vector. header}

\item{cle_var}{string. key column name}

\item{classe_var}{string. class column name}

\item{format_fun}{list. formatting functions}
}
\description{
Print tasks to latex
}

