% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.r
\name{print_taches}
\alias{print_taches}
\title{Print tasks to latex}
\usage{
print_taches(taches, legende, col_format, rowcol_head = "lightgray",
  header = c("TACHE", "ECHEANCE", "ETAT"), format_fun, ...)
}
\arguments{
\item{taches}{data.frame}

\item{legende}{data.frame}

\item{col_format}{string. column format}

\item{rowcol_head}{row color of the header}

\item{header}{character vector. header}

\item{format_fun}{list. formatting functions}

\item{...}{further arguments to be passed to print_table}
}
\description{
Print tasks to latex
}

