% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xl.r
\name{read_xl}
\alias{read_xl}
\title{Read excel file}
\usage{
read_xl(xl_file, sheets = c("LEGENDE", "TACHES", "CEJOUR", "PLANS",
  "PLANSNOTE"), col_dates = c("DATE", "ECHEANCE", "DATEREALISATION"),
  origin = "1899-12-30")
}
\arguments{
\item{xl_file}{string. path to the file}

\item{sheets}{character or integer vector. mandatory sheets to be read}

\item{col_dates}{character vector. date column names}

\item{origin}{a Date object, or something which can be coerced by
  \code{as.Date(origin)} to such an object.}
}
\description{
Read excel file
}

