% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/export.r
\name{resize_photos}
\alias{resize_photos}
\title{Resize photos}
\usage{
resize_photos(photo_files = list.files(".", pattern =
  ".*\\\\.(jpg|jpeg|JPG|JPEG|png|PNG)"), folder = "tmp", max_width = 800,
  max_height = 600)
}
\arguments{
\item{photo_files}{character vector. paths of the photo files}

\item{folder}{string. output folder of the resized photos}

\item{max_width}{maximum photo width}

\item{max_height}{maximum photo height}
}
\description{
Resize photos
}

