% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xl.r
\name{write_xl}
\alias{write_xl}
\title{Write data.frame to Excel file}
\usage{
write_xl(df, out_file, open = FALSE, zip_path = "C:\\\\Rtools\\\\bin", ...)
}
\arguments{
\item{df}{data.frame or list of data.frames}

\item{out_file}{string. path to the output excel file}

\item{open}{logical. activate opening the file in Excel/LibreOffice/OpenOffice}

\item{zip_path}{string. path containing zip executable on windows}

\item{...}{further arguments to pass to \code{\link[openxlsx]{write.xlsx}}}
}
\description{
Write data.frame to Excel file
}

